requirements.txt
amqp==5.1.1
bcrypt==4.0.1
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Celery==2.4.3
Flask-Script==2.0.6
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
prompt-toolkit==3.0.36
psycopg2==2.9.5
PyJWT==2.6.0
python-dotenv==0.21.0
pytz==2022.7
six==1.16.0
SQLAlchemy==1.4.45
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.2.2
